+++ example/a.ts
@@ -1,8 +1,30 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
-const app = new Elysia().get('/สวัสดี', 'สบายดีไหม').listen(0)
+const app = new Elysia()
+	.wrap((fn, request) => {
+		const _request = request.clone()
 
-const response = await fetch(`http://localhost:${app.server!.port}/สวัสดี`)
-const text = await response.text()
+		return () => {
+			try {
+				return fn(request.clone())
+			} catch {
+				console.log('ER')
+			}
+		}
+	})
+	.onError(({ error }) => {
+		if (error) throw error
+	})
+	.get('/', () => {
+		throw new Error('A')
 
-console.log(text)
+		return 'Hello World!'
+	})
+	.listen(3000)
+
+// console.log(app.fetch.toString())
+
+app.handle(req('/'))
+	.then((x) => x.text())
+	.then(console.log)