+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 Change:
 - do not prettify routes when using `guard`, `group`
 - use `process.getBuiltinModule` instead of dynamic import for file
+- experimental `adapter/cloudflare-worker`
+- add `ElysiaAdapter.beforeCompile`
 
 # 1.4.6 - 18 Sep 2025
 Improvement: